backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
openai>=1.35.0
pgvector>=0.2.4
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.7.4,<3
httpx>=0.27.0
slowapi>=0.1.9
